# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CalibrationRecordFormType =
  (typeof CalibrationRecordFormType)[keyof typeof CalibrationRecordFormType];

export const CalibrationRecordFormType = {
  control_valve: "control_valve",
  timbangan: "timbangan",
  rtd: "rtd",
  ph: "ph",
  transmitter: "transmitter",
} as const;

/**
 * JSON object containing all form-specific data
 */
export type CalibrationRecordData = { [key: string]: unknown };

export interface CalibrationRecord {
  id: number;
  formType: CalibrationRecordFormType;
  tagNo?: string;
  date?: string;
  calibratedBy?: string;
  /** JSON object containing all form-specific data */
  data: CalibrationRecordData;
  createdAt: string;
  updatedAt: string;
}

export type CreateCalibrationBodyFormType =
  (typeof CreateCalibrationBodyFormType)[keyof typeof CreateCalibrationBodyFormType];

export const CreateCalibrationBodyFormType = {
  control_valve: "control_valve",
  timbangan: "timbangan",
  rtd: "rtd",
  ph: "ph",
  transmitter: "transmitter",
} as const;

export type CreateCalibrationBodyData = { [key: string]: unknown };

export interface CreateCalibrationBody {
  formType: CreateCalibrationBodyFormType;
  tagNo?: string;
  date?: string;
  calibratedBy?: string;
  data: CreateCalibrationBodyData;
}

export type CalibrationSummaryByType = {
  control_valve: number;
  timbangan: number;
  rtd: number;
  ph: number;
  transmitter: number;
};

export interface CalibrationSummary {
  total: number;
  byType: CalibrationSummaryByType;
}

export type ListCalibrationsParams = {
  formType?: ListCalibrationsFormType;
  tagNo?: string;
};

export type ListCalibrationsFormType =
  (typeof ListCalibrationsFormType)[keyof typeof ListCalibrationsFormType];

export const ListCalibrationsFormType = {
  control_valve: "control_valve",
  timbangan: "timbangan",
  rtd: "rtd",
  ph: "ph",
  transmitter: "transmitter",
} as const;