# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all calibration records
 */
export const ListCalibrationsQueryParams = zod.object({
  formType: zod
    .enum(["control_valve", "timbangan", "rtd", "ph", "transmitter"])
    .optional(),
  tagNo: zod.coerce.string().optional(),
});

export const ListCalibrationsResponseItem = zod.object({
  id: zod.number(),
  formType: zod.enum([
    "control_valve",
    "timbangan",
    "rtd",
    "ph",
    "transmitter",
  ]),
  tagNo: zod.string().optional(),
  date: zod.string().optional(),
  calibratedBy: zod.string().optional(),
  userId: zod.number().nullable().optional(),
  data: zod
    .record(zod.string(), zod.unknown())
    .describe("JSON object containing all form-specific data"),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListCalibrationsResponse = zod.array(ListCalibrationsResponseItem);

/**
 * @summary Create calibration record
 */
export const CreateCalibrationBody = zod.object({
  formType: zod.enum([
    "control_valve",
    "timbangan",
    "rtd",
    "ph",
    "transmitter",
  ]),
  tagNo: zod.string().optional(),
  date: zod.string().optional(),
  calibratedBy: zod.string().optional(),
  data: zod.record(zod.string(), zod.unknown()),
});

/**
 * @summary Get calibration record by ID
 */
export const GetCalibrationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCalibrationResponse = zod.object({
  id: zod.number(),
  formType: zod.enum([
    "control_valve",
    "timbangan",
    "rtd",
    "ph",
    "transmitter",
  ]),
  tagNo: zod.string().optional(),
  date: zod.string().optional(),
  calibratedBy: zod.string().optional(),
  userId: zod.number().nullable().optional(),
  data: zod
    .record(zod.string(), zod.unknown())
    .describe("JSON object containing all form-specific data"),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update calibration record
 */
export const UpdateCalibrationParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCalibrationBody = zod.object({
  formType: zod.enum([
    "control_valve",
    "timbangan",
    "rtd",
    "ph",
    "transmitter",
  ]),
  tagNo: zod.string().optional(),
  date: zod.string().optional(),
  calibratedBy: zod.string().optional(),
  data: zod.record(zod.string(), zod.unknown()),
});

export const UpdateCalibrationResponse = zod.object({
  id: zod.number(),
  formType: zod.enum([
    "control_valve",
    "timbangan",
    "rtd",
    "ph",
    "transmitter",
  ]),
  tagNo: zod.string().optional(),
  date: zod.string().optional(),
  calibratedBy: zod.string().optional(),
  userId: zod.number().nullable().optional(),
  data: zod
    .record(zod.string(), zod.unknown())
    .describe("JSON object containing all form-specific data"),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete calibration record
 */
export const DeleteCalibrationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get summary counts per form type
 */
export const GetCalibrationSummaryResponse = zod.object({
  total: zod.number(),
  byType: zod.object({
    control_valve: zod.number(),
    timbangan: zod.number(),
    rtd: zod.number(),
    ph: zod.number(),
    transmitter: zod.number(),
  }),
});