# lib/api-zod/src/generated/types/calibrationRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalibrationRecordData } from "./calibrationRecordData";
import type { CalibrationRecordFormType } from "./calibrationRecordFormType";

export interface CalibrationRecord {
  id: number;
  formType: CalibrationRecordFormType;
  tagNo?: string;
  date?: string;
  calibratedBy?: string;
  /** JSON object containing all form-specific data */
  data: CalibrationRecordData;
  createdAt: string;
  updatedAt: string;
}