# lib/api-zod/src/generated/types/calibrationSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalibrationSummaryByType } from "./calibrationSummaryByType";

export interface CalibrationSummary {
  total: number;
  byType: CalibrationSummaryByType;
}