# lib/api-zod/src/generated/types/createCalibrationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCalibrationBodyData } from "./createCalibrationBodyData";
import type { CreateCalibrationBodyFormType } from "./createCalibrationBodyFormType";

export interface CreateCalibrationBody {
  formType: CreateCalibrationBodyFormType;
  tagNo?: string;
  date?: string;
  calibratedBy?: string;
  data: CreateCalibrationBodyData;
}